Change integer division to multiply + shift.
The magic number 0x1012 with right shift 12 was chosen experimentally
and Just Works™. It introduces just one error for all the possible
result values: for target alpha range 22, when the source range is 241,
the resulting value is 12, instead of 11.

There is a no-errors magic value, 0x201F with right shift 13, but it
doesn't fit into uint16_t number:

Maximum source range ------.
Minimum target range -.    |
                      |    |
                      v    v
            0x201F / 18 * 256 = 0x1C800

While the 0x1012 version fits:

            0x1012 / 18 * 256 = 0xE400

This will be important for the SIMD version of the algorithm. Since we
want to have deterministic results on both versions of code, we also
introduce this error in the reference implementation.

This change has no visible performance impact.