Prevent deadlock on TaskDispatch destruction.
If a worker thread executes a predicate check on m_cvWork.wait() while
the dispatcher signals the condition variable, the worker won't get
the notification and will sleep indefinitely. This causes
the dispatcher to hand on worker.join() further below.

Locking the notify_all() call ensures that every worker sleeps
on the condvar and thus receives the signal.

(This fix was originally discovered by Bartosz Szreder.)