Fix handling of pthread naming API for LLVM on Windows
The previous two-step method relying on `std:thread:native_handle()` is tricky
as it's implementation-defined, and it seems that LLVM's libc++ does not return
a `pthread_t` handle even when using pthread (and the MSVC code path is not an
option for LLVM on Windows).

With mingw-gcc as packaged in Linux distros with pthread support it worked
fine, but with llvm-mingw it was problematic.
It seems like it could also be problematic with clang-cl.

Setting the name in the thread directly as done for Apple platforms is simpler
and should work fine.

Co-authored-by: Hein-Pieter van Braam-Stewart <hp@tmm.cx>