Remove support for alpha-as-rgb processing.
ETC1 did not support RGBA textures, so the usual hack for handling RGBA
data was to create two ETC1 textures, one with RGB data, and the second
one with alpha data, then combine them in a shader.

ETC2 is now ubiquitous, so support for the hacks of old can go away.